+++ src/collider.js
@@ -4,16 +4,6 @@
     this._getCollisionPairs = quadTreeCollisionPairs;
   };
 
-  // if no entities have uncollision(), skip expensive record keeping for uncollisions
-  var isUncollisionOn = function(entities) {
-    for (var i = 0, len = entities.length; i < len; i++) {
-      if (entities[i].uncollision !== undefined) {
-        return true;
-      }
-    }
-    return false;
-  };
-
   var isSetupForCollisions = function(obj) {
     return obj.center !== undefined && obj.size !== undefined;
   };
@@ -79,18 +69,12 @@
     },
 
     collision: function(entity1, entity2) {
-      var collisionType;
-      if (!isUncollisionOn(this.c.entities.all())) {
-        collisionType = this.INITIAL;
-      } else if (this.getCollideRecordIds(entity1, entity2).length === 0) {
+      if (this.getCollideRecordIds(entity1, entity2).length === 0) {
         this._collideRecords.push([entity1, entity2]);
-        collisionType = this.INITIAL;
-      } else {
-        collisionType = this.SUSTAINED;
       }
 
-      notifyEntityOfCollision(entity1, entity2, collisionType);
-      notifyEntityOfCollision(entity2, entity1, collisionType);
+      notifyEntityOfCollision(entity1, entity2);
+      notifyEntityOfCollision(entity2, entity1);
     },
 
     createEntity: function(entity) {
@@ -103,27 +87,12 @@
     },
 
     destroyEntity: function(entity) {
-      var recordIds = this.getCollideRecordIds(entity);
-      for (var i = 0; i < recordIds.length; i++) {
-        this.removeOldCollision(recordIds[i]);
-      }
-
       // if coll detection happening, remove any pairs that include entity
       for(var i = this._currentCollisionPairs.length - 1; i >= 0; i--){
         if (this._currentCollisionPairs[i][0] === entity ||
            this._currentCollisionPairs[i][1] === entity) {
           this._currentCollisionPairs.splice(i, 1);
         }
-      }
-    },
-
-    // remove collision at passed index
-    removeOldCollision: function(recordId) {
-      var record = this._collideRecords[recordId];
-      if (record !== undefined) {
-        notifyEntityOfUncollision(record[0], record[1])
-        notifyEntityOfUncollision(record[1], record[0])
-        this._collideRecords.splice(recordId, 1);
       }
     },
 
@@ -150,8 +119,9 @@
       }
     },
 
-    INITIAL: 0,
-    SUSTAINED: 1,
+    isIntersecting: function(obj1, obj2) {
+      isIntersecting(obj1, obj2);
+    },
 
     RECTANGLE: 0,
     CIRCLE: 1
@@ -220,18 +190,12 @@
   };
 
   var getBoundingBox = function(obj) {
-    return obj.boundingBox || new Coquette.Collider.Shape.Rectangle(obj);
+    return obj.boundingBox || new Collider.Shape.Rectangle(obj);
   };
 
-  var notifyEntityOfCollision = function(entity, other, type) {
+  var notifyEntityOfCollision = function(entity, other) {
     if (entity.collision !== undefined) {
-      entity.collision(other, type);
-    }
-  };
-
-  var notifyEntityOfUncollision = function(entity, other) {
-    if (entity.uncollision !== undefined) {
-      entity.uncollision(other);
+      entity.collision(other);
     }
   };